requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
Pillow==8.3.2
tensorflow==2.6.0